feat(user-selection): allow jumping to user by typing id
This allows typing a user id and then pressing enter to jump directly to
a user from the user selection screen (active or inactive).

The use case serving as motivation for this enhancement is scanning a
Code 39 barcode, which is only required to contain the user id and
start-/stop characters. This very short barcode can be 3d-printed on
mate tags and enables users to buy something using only the barcode
scanner. This way, they don't have to touch the screen, which in
combination with a stationary scanner allows for a completely touchless
interface.

Closely related to #111, but with smaller space required for the
barcode.